+++ backend/database/migrations/2023_08_05_184952_create_snapshot_rkarinciankegiatan_table.php
@@ -50,7 +50,7 @@ class CreateSnapshotRkarinciankegiatanTable extends Migration
       $table->tinyInteger('EntryLvl')->default(0);
       $table->string('Descr')->nullable();            
       $table->year('TA'); 
-      $table->tinyInteger('TABULAN');
+      $table->smallInteger('TABULAN');
       $table->boolean('Locked')->default(0);
       $table->uuid('RKARincID_Src')->nullable();
       $table->timestamps();
