+++ backend/database/migrations/2023_08_05_190813_create_snapshot_rkatargetrinciankegiatan_table.php
@@ -26,7 +26,7 @@ class CreateSnapshotRkatargetrinciankegiatanTable extends Migration
       $table->tinyInteger('EntryLvl')->default(0);
       $table->string('Descr')->nullable();            
       $table->year('TA'); 
-      $table->tinyInteger('TABULAN');
+      $table->smallInteger('TABULAN');
       $table->boolean('Locked')->default(0);
       $table->uuid('RKATargetRincID_Src')->nullable();
       $table->timestamps();
